+++ gDRutils/R/assay_names.R
@@ -1,0 +1,147 @@
+ASSAY_INFO_TBL <- tibble::tribble(
+  ~ name, ~pname, ~ type, ~ group, ~ data_type,
+  "rawTreated", "Raw Treated", "raw", "raw", "single-agent",
+  "Controls", "Controls", "control", "core", "single-agent",
+  "Normalized", "Normalized", "normalized", "core", "single-agent",
+  "Averaged", "Averaged", "averaged", "core", "single-agent",
+  "Metrics", "Metrics", "metrics", "core", "single-agent",
+  "SmoothMatrix", "MX full", "smooth", "combo_base_mx", "combo",
+  "BlissExcess", "Bliss excess", "bliss_excess", "combo_base_excess", "combo",
+  "HSAExcess", "HSA excess", "hsa_excess", "combo_base_excess", "combo",
+  "HSAScore", "HSA excess", "hsa_score", "combo_score_excess", "combo",
+  "BlissScore", "Bliss excess", "bliss_score", "combo_score_excess", "combo",
+  "CIScore_50", "log2(CI) @ GR/IC50", "ci_score_50", "combo_score_mx", "combo",
+  "CIScore_80", "log2(CI) @ GR/IC80", "ci_score_80", "combo_score_mx", "combo",
+  "isobolograms", "isobolograms", "iso", "combo_iso", "combo"
+)
+
+
+#' get default assay names
+#' for the specified filters,
+#' i.e. set of assay types, assay groups and assay data types
+#'
+#' @param type charvec assay types
+#' @param group charvec assay grups
+#' @param data_type charvec assay data types
+#' @param prettify logical flag, prettify the assay name?
+#' @param simplify logical flag, simplify the output?
+#'    will return single string instead of named vector with single element
+#'    useful when function is expect to return single element/assay only
+#'
+#' @author Arkadiusz Gładki \email{arkadiusz.gladki@@contractors.roche.com}
+#'
+get_env_assay_names <-
+  function(type = NULL,
+           group = NULL,
+           data_type = NULL,
+           prettify = FALSE,
+           simplify = TRUE) {
+    assert_choices(type, choices = ASSAY_INFO_TBL$type, null.ok = TRUE)
+    assert_choices(group, choices = ASSAY_INFO_TBL$group, null.ok = TRUE)
+    assert_choices(data_type,
+                   choices = ASSAY_INFO_TBL$data_type,
+                   null.ok = TRUE)
+    checkmate::assert_flag(prettify)
+
+    # to prettify or not to prettify
+    # this is the question
+    fname <- if (prettify) {
+      "pname"
+    } else {
+      "name"
+    }
+
+    filters <-
+      list(type = type,
+           group = group,
+           data_type = data_type)
+    are_null <-
+      vapply(filters, is.null, logical(1))
+    v_filters <- filters[!are_null]
+
+    if (all(are_null)) {
+      return(structure(ASSAY_INFO_TBL[[fname]], names = ASSAY_INFO_TBL$type))
+    }
+      
+    df <- ASSAY_INFO_TBL
+    for (filter in names(v_filters)) {
+      df <- df[df[[eval(filter)]] %in% v_filters[[filter]], ]
+    }
+
+    if ((nrow(df)) == 0)  {
+      v_filters_str <-
+        paste(names(v_filters),
+              v_filters,
+              sep = "=",
+              collapse = "&")
+      msg <- sprintf("Assay name not found for '%s'", v_filters_str)
+      stop(msg)
+    }
+
+    res <- structure(df[[fname]], names = df$type)
+    if (length(res) == 1 && simplify == TRUE) {
+      as.character(res)
+    } else {
+      res
+    }
+  }
+
+#' get assay names of the given se/dataset
+#' fetch the data from the se if provided as metadata
+#' use predefined values from `get_env_assay_names` otherwise
+#'
+#' @param se SummarizedExperiment or NULL
+#' 
+#' @inheritParams get_env_assay_names
+#'
+#' @author Arkadiusz Gładki \email{arkadiusz.gladki@@contractors.roche.com}
+#'
+#' @export
+get_assay_names <- function(se = NULL, ...) {
+  if (!is.null(se) &&
+      !is.null(metadata(se = NULL)[["assay_info_tbl"]])) {
+    # TODO: extend the logic to support metadata from se
+    # i.e. sth like `get_se_assay_names(group = "combo_base")`
+    # more details: https://jira.gene.com/jira/browse/GDR-1116
+    stop("'get_assay_name' currently does not support non-null 'se'")
+  } else {
+    get_env_assay_names(...)
+  }
+}
+
+#' get names of combo assays
+#' @return  charvec
+#' @export
+#'
+#' @author Arkadiusz Gładki \email{arkadiusz.gladki@@contractors.roche.com}
+#'
+get_combo_assay_names <- function(se = NULL, ...) {
+        get_assay_names(se, data_type = "combo", simplify = FALSE, ...)
+}
+
+#' get names of combo base assays
+#' 
+#' @param se SummarizedExperiment or NULL
+#' 
+#' @return  charvec
+#' @export
+#'
+#' @author Arkadiusz Gładki \email{arkadiusz.gladki@@contractors.roche.com}
+#'
+get_combo_base_assay_names <- function(se = NULL, ...) {
+        get_combo_assay_names(group = c("combo_base_excess", "combo_base_mx"), ...)
+}
+
+#' get names of combo score assays
+#' 
+#' @param se SummarizedExperiment or NULL
+#' 
+#' @return  charvec
+#' 
+#' @export
+#'
+#' @author Arkadiusz Gładki \email{arkadiusz.gladki@@contractors.roche.com}
+#'
+get_combo_score_assay_names <- function(se = NULL, ...) {
+        get_combo_assay_names(group = c("combo_score_excess", "combo_score_mx"), ...)
+}