+++ gDRutils/R/utils.R
@@ -25,3 +25,46 @@ assert_equal_input_len <- function(outlier, ...) {
 
   invisible(NULL)
 }
+
+#' shorten normalization type
+#' 
+#' @param x string with normalization type
+#' 
+#' @export
+shorten_normalization_type_name <- function(x) {
+  checkmate::assert_choice(x, c("RelativeViability", "GRvalue"))
+  dict <- c("RelativeViability" = "RV", "GRvalue" = "GR")
+  dict[[x]]
+}
+
+#' extend abbreviated normalization type
+#' 
+#' @param x string with normalization type
+#' 
+#' @export
+extend_normalization_type_name <- function(x) {
+  checkmate::assert_choice(x, c("RV", "GR"))
+  dict <- c("RV" = "RelativeViability", "GR" = "GRvalue")
+  dict[[x]]
+}
+
+#' assert choices
+#' 
+#' @param x charvec expected subset
+#' @param choices charvec reference set
+#' @export
+assert_choices <- function(x, choices, ...) {
+  out <- vapply(x, function(y) {
+    checkmate::test_choice(y, choices, ...)
+  }, FUN.VALUE = logical(1))
+
+  if (!all(out)) {
+    msg <-
+      sprintf(
+        "Assertion on '%s' failed. Must be element(s) of {'%s'} set.",
+        toString(x[!out]),
+        toString(choices)
+      )
+    stop(msg)
+  }
+}